requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
tqdm
numpy
torch
google-generativeai
python-dotenv
